+++ packages/waku/src/cli.ts
@@ -56,10 +56,9 @@ if (values.version) {
 
 async function runDev(options: { ssr: boolean }) {
   const { Hono } = await import('hono');
-  const { honoWrapper } = await import('./lib/middleware/honoWrapper.js');
-  const { rsc } = await import('./lib/middleware/rsc.js');
+  const { honoMiddleware } = await import('./lib/middleware/hono.js');
   const app = new Hono();
-  app.use('*', honoWrapper(rsc({ config, command: 'dev', ssr: options.ssr })));
+  app.use('*', honoMiddleware({ config, command: 'dev', ssr: options.ssr }));
   const port = parseInt(process.env.PORT || '3000', 10);
   startServer(app, port);
 }
@@ -73,13 +72,9 @@ async function runStart(options: { ssr: boolean }) {
   const { Hono } = await import('hono');
   const { serveStatic } = await import('@hono/node-server/serve-static');
   const { resolveConfig } = await import('./lib/config.js');
-  const { honoWrapper } = await import('./lib/middleware/honoWrapper.js');
-  const { rsc } = await import('./lib/middleware/rsc.js');
+  const { honoMiddleware } = await import('./lib/middleware/hono.js');
   const app = new Hono();
-  app.use(
-    '*',
-    honoWrapper(rsc({ config, command: 'start', ssr: options.ssr })),
-  );
+  app.use('*', honoMiddleware({ config, command: 'start', ssr: options.ssr }));
   const { rootDir, distDir, publicDir } = await resolveConfig(config);
   app.use(
     '*',
