+++ packages/waku/src/client.ts
@@ -13,7 +13,7 @@ import {
 } from 'react';
 import type { ReactNode } from 'react';
 import RSDWClient from 'react-server-dom-webpack/client';
-import { encodeInput } from './lib/middleware/rsc/utils.js';
+import { encodeInput } from './lib/rsc/utils.js';
 
 const { createFromFetch, encodeReply } = RSDWClient;
