+++ packages/waku/src/lib/rsc/worker-impl.ts
@@ -6,13 +6,13 @@ import { Server } from 'node:http';
 import { createServer as viteCreateServer } from 'vite';
 import type { ViteDevServer } from 'vite';
 
-import { fileURLToFilePath } from '../../utils/path.js';
+import { fileURLToFilePath } from '../utils/path.js';
 import { hasStatusCode, deepFreeze } from './utils.js';
 import type { MessageReq, MessageRes, RenderRequest } from './worker-api.js';
-import { renderRSC } from '../../rsc/renderer.js';
-import { rscTransformPlugin } from '../../vite-plugin/rsc-transform-plugin.js';
-import { rscReloadPlugin } from '../../vite-plugin/rsc-reload-plugin.js';
-import { rscDelegatePlugin } from '../../vite-plugin/rsc-delegate-plugin.js';
+import { renderRSC } from './renderer.js';
+import { rscTransformPlugin } from '../vite-plugin/rsc-transform-plugin.js';
+import { rscReloadPlugin } from '../vite-plugin/rsc-reload-plugin.js';
+import { rscDelegatePlugin } from '../vite-plugin/rsc-delegate-plugin.js';
 
 const IS_NODE_20 = Number(process.versions.node.split('.')[0]) >= 20;
 if (IS_NODE_20) {
