+++ packages/waku/src/lib/vite-plugin/rsc-index-plugin.ts
@@ -1,7 +1,6 @@
 import type { Plugin } from 'vite';
 
-// FIXME we should avoid external dependencies out of this file.
-import { codeToInject } from '../middleware/rsc/utils.js';
+import { codeToInject } from '../rsc/utils.js';
 
 export function rscIndexPlugin(cssAssets: string[]): Plugin {
   return {
