+++ packages/waku/src/main.ts
@@ -1,6 +1,6 @@
-export { honoWrapper } from './lib/middleware/honoWrapper.js';
-export { connectWrapper } from './lib/middleware/connectWrapper.js';
-export { rsc } from './lib/middleware/rsc.js';
+export { honoMiddleware } from './lib/middleware/hono.js';
+export { connectMiddleware } from './lib/middleware/connect.js';
+export { createHandler } from './lib/rsc/handler.js';
 
 import type { build as buildOrig } from './lib/builder.js';
